fix(inference): stop 64-bit atomics from breaking every Cortex-M link
BLOCKER: __atomic_store_n/__atomic_load_n on a uint64_t
last_invoke_latency_us lowered to __atomic_store_8/__atomic_load_8 on
Cortex-M (no HW 64-bit atomic, and the Zephyr SDK's arm-zephyr-eabi
toolchain ships no libatomic), so every app linking
src/inference_dispatch.c (compiled unconditionally by
zephyr/CMakeLists.txt) that actually called the new accessor's
invoke/latency path would fail to link. Narrow the Zephyr-side field to
a uint32_t (naturally atomic on Cortex-M, no libcall); the public
out_us stays uint64_t. Ceiling: saturates at UINT32_MAX us (~71.58
min); documented on the field and in the header. The Yocto dispatcher
(x86-64/aarch64, native 64-bit atomics) is untouched.

MAJOR: alp_inference_invoke() now uses k_cycle_get_64() on
CONFIG_TIMER_HAS_64BIT_CYCLE_COUNTER targets (e.g. the AEN801 M55,
where a plain k_cycle_get_32() wraps every ~10.74s at 400MHz) instead
of always using the 32-bit counter, and saturates rather than silently
wraps if the measured duration would exceed the 32-bit storage field.

MAJOR: alp_inference_last_invoke_latency_us() now checks out_us for
NULL before the handle on BOTH dispatchers (Zephyr already did; Yocto
checked the handle first), so (NULL, NULL) and (closed, NULL) return
ALP_ERR_INVAL consistently instead of disagreeing by OS. Covered by new
tests on both dispatchers.

MAJOR: document the stub build's ALP_ERR_NOSUPPORT return in the
header -- previously an undocumented status for this accessor.

Also: document last-store-wins semantics for overlapping invokes on one
handle, the microseconds-vs-milliseconds unit note for the tier-2
recipe schema, and add a deterministic invoke-vs-invoke overlap race
test (tests/yocto/inference_invoke_overlap_race.c) since nothing
exercised the concurrent-writer path this design relies on. Strengthen
the Zephyr unit test's fake_invoke_ok to do real measurable work so the
populated-after-invoke assertion can't pass against a hardcoded-0
latency. Regenerate tests/fixtures/stub-symbol-matrix/symbols.json
(reviewed diff: the new symbol added to the 29 combos that carry
inference stubs at all).

+++ src/backends/inference/inference_ops.h
@@ -82,21 +82,43 @@ struct alp_inference {
 	alp_capabilities_t            cached_caps;
 	/* Last successful alp_inference_invoke()'s wall-clock duration, in
 	 * microseconds -- written by the dispatcher itself (it brackets
-	 * ops->invoke), never by a backend.  UINT64_MAX means "no
-	 * successful invoke yet"; alp_inference_last_invoke_latency_us()
+	 * ops->invoke), never by a backend.  uint32_t, NOT uint64_t
+	 * (issue: 64-bit atomics don't link on Cortex-M -- M-profile has
+	 * no 64-bit atomic instruction, GCC lowers __atomic_{load,store}_n
+	 * on a uint64_t to __atomic_load_8/__atomic_store_8 libatomic
+	 * calls, and the Zephyr SDK's arm-zephyr-eabi toolchain ships no
+	 * libatomic, so every Cortex-M app failed to link the moment this
+	 * dispatcher -- which zephyr/CMakeLists.txt compiles
+	 * unconditionally -- pulled one in). A naturally-aligned 32-bit
+	 * load/store IS lock-free on every Cortex-M this SDK targets (a
+	 * single LDR/STR, exactly like active_ops/lifecycle below), so
+	 * __atomic_store_n/__atomic_load_n on this field compiles to plain
+	 * instructions with no libcall. The tradeoff is range: values
+	 * saturate at UINT32_MAX us (~4294.967295 s, ~71.58 minutes) --
+	 * see alp_inference_last_invoke_latency_us()'s header doc
+	 * (<alp/inference.h>) for the ceiling and the saturate-not-wrap
+	 * behaviour. UINT32_MAX doubles as the "no successful invoke yet"
+	 * sentinel (a real invoke can plausibly hit UINT32_MAX-1 us but
+	 * hitting the exact all-ones value and being misread as
+	 * NOT_READY is not worth guarding -- a wall-clock invoke pinned to
+	 * that specific microsecond is not a value one is exactly on the
+	 * ceiling); alp_inference_last_invoke_latency_us()
 	 * (src/inference_dispatch.c) translates that sentinel to
 	 * ALP_ERR_NOT_READY at the public API boundary rather than
 	 * exposing it.  The claim-time memset in _alloc() zeroes this
 	 * field along with everything else before in_use, so
-	 * alp_inference_open() must explicitly (re)set it to UINT64_MAX
+	 * alp_inference_open() must explicitly (re)set it to UINT32_MAX
 	 * once the handle is claimed -- a fresh handle must read
 	 * NOT_READY, not "0 us". active_ops permits concurrent ops on one
 	 * handle (it is a drain counter, not a mutex -- see
 	 * alp_slot_claim.h), so a concurrent invoke()-writer and
 	 * latency-reader is a real interleaving: always access this field
 	 * via __atomic_store_n/__atomic_load_n, never a plain read/write
-	 * (issue #629 discipline). */
-	uint64_t last_invoke_latency_us;
+	 * (issue #629 discipline). The public accessor's out_us stays
+	 * uint64_t (see <alp/inference.h>) -- only the internal storage
+	 * narrowed, so this is not an ABI change to the new (still
+	 * ABI-EXPERIMENTAL) accessor's signature. */
+	uint32_t last_invoke_latency_us;
 	/* lifecycle/active_ops drive the generic open/op/close guard in
 	 * src/common/alp_slot_claim.h (alp_handle_op_enter/leave/
 	 * begin_close, issue #629) -- placed before in_use so the atomic-
